hoàn thiện toàn bộ hệ thống bán hàng và tích hợp thanh toán VNPAY

+++ WebBanHangOnline/Models/ShoppingCart.cs
@@ -12,10 +12,11 @@ namespace WebBanHangOnline.Models
         {
             this.Items = new List<ShoppingCartItem>();
         }
-        public void AddToCart(ShoppingCartItem item, int Quantity)
+
+        public void AddToCart(ShoppingCartItem item,int Quantity)
         {
             var checkExits = Items.FirstOrDefault(x => x.ProductId == item.ProductId);
-            if(checkExits != null)
+            if (checkExits != null)
             {
                 checkExits.Quantity += Quantity;
                 checkExits.TotalPrice = checkExits.Price * checkExits.Quantity;
@@ -25,16 +26,17 @@ namespace WebBanHangOnline.Models
                 Items.Add(item);
             }
         }
+
         public void Remove(int id)
         {
-            var checkExits = Items.SingleOrDefault(x => x.ProductId == id);
-            if(checkExits != null)
+            var checkExits = Items.SingleOrDefault(x=>x.ProductId==id);
+            if (checkExits != null)
             {
                 Items.Remove(checkExits);
             }
         }
 
-        public void UpdateQuantity(int id, int quantity)
+        public void UpdateQuantity(int id,int quantity)
         {
             var checkExits = Items.SingleOrDefault(x => x.ProductId == id);
             if (checkExits != null)
@@ -43,12 +45,12 @@ namespace WebBanHangOnline.Models
                 checkExits.TotalPrice = checkExits.Price * checkExits.Quantity;
             }
         }
+
         public decimal GetTotalPrice()
         {
-            return Items.Sum(x => x.TotalPrice);
+            return Items.Sum(x=>x.TotalPrice);
         }
-
-        public decimal GetTotalQuantity()
+        public int GetTotalQuantity()
         {
             return Items.Sum(x => x.Quantity);
         }
@@ -56,7 +58,9 @@ namespace WebBanHangOnline.Models
         {
             Items.Clear();
         }
+
     }
+
     public class ShoppingCartItem
     {
         public int ProductId { get; set; }
@@ -68,5 +72,4 @@ namespace WebBanHangOnline.Models
         public decimal Price { get; set; }
         public decimal TotalPrice { get; set; }
     }
-
 }
